api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
asyncpg>=0.29.0
sqlalchemy[asyncio]>=2.0.0
pydantic>=2.5.0
pydantic-settings>=2.0.0
openai>=1.3.0
celery>=5.3.0
redis>=5.0.0
sendgrid>=6.10.0
python-dotenv>=1.0.0
numpy>=1.26.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Enhanced authentication
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Monitoring and health
psutil>=5.9.0

# Vector database and AI
chromadb>=0.4.15
langchain>=0.1.0
langchain-community>=0.0.10
sentence-transformers>=2.2.2
huggingface-hub>=0.19.0

# HTTP client
httpx>=0.25.0

# File handling
python-magic>=0.4.27
PyPDF2>=3.0.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Logging
structlog>=23.2.0 